augment-code-free-gui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "augment-free"
version = "0.1.2"
description = "Free AugmentCode - A tool for cleaning AugmentCode related data with modern GUI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "vagmr"}
]
keywords = ["augmentcode", "vscode", "telemetry", "cleaner", "gui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "pywebview>=4.4.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
augment-free = "augment_free.main:main"

[project.urls]
Homepage = "https://github.com/vagmr/Augment-free"
Repository = "https://github.com/vagmr/Augment-free"
Issues = "https://github.com/vagmr/Augment-free/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/augment_free"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
